Add a list of frequently used terms and their translations
Add a list of terms frequently used in the A+ user interface or in
relation to A+ to support further improvement of UI texts as well as
help teachers and material creaters to use consistent terminology.

The list is in .rst-format and might not be supported by Jekyll and
Github pages. However, the table and text is formatted and can be read
on Github.