Fix the contributors script: empty repos and anon profiles
Empty git repos caused crashes: they are skipped now.

There was a case where the GitHub user profile link was not added
for a user that has added their real name to the GitHub profile.
If the email address in the git commit is not connected to any
GitHub user, then the profile link was excluded in the script.
This has been fixed. It is typical that people remove
university email addresses from their GitHub profiles
once they lose access to the university email.

In addition, update the pip packages in
update_contributors_requirements.txt.